+++ lib/plausible/billing/qouta/quota.ex
@@ -4,30 +4,8 @@ defmodule Plausible.Billing.Quota do
   """
 
   use Plausible
-  alias Plausible.Users
-  alias Plausible.Auth.User
-  alias Plausible.Billing.{Plan, Plans, EnterprisePlan}
+  alias Plausible.Billing.{Plan, EnterprisePlan}
   alias Plausible.Billing.Quota.{Usage, Limits}
-
-  @doc """
-  Enterprise plans are always allowed to add more sites (even when
-  over limit) to avoid service disruption. Their usage is checked
-  in a background job instead (see `check_usage.ex`).
-  """
-  def ensure_can_add_new_site(user) do
-    user = Users.with_subscription(user)
-
-    case Plans.get_subscription_plan(user.subscription) do
-      %EnterprisePlan{} ->
-        :ok
-
-      _ ->
-        usage = Usage.site_usage(user)
-        limit = Limits.site_limit(user)
-
-        if below_limit?(usage, limit), do: :ok, else: {:error, {:over_limit, limit}}
-    end
-  end
 
   @doc """
   Ensures that the given user (or the usage map) is within the limits
@@ -37,14 +15,10 @@ defmodule Plausible.Billing.Quota do
   which bypasses the pageview limit check and returns `:ok` as long as
   the other limits are not exceeded.
   """
-  @spec ensure_within_plan_limits(User.t() | map(), struct() | atom() | nil, Keyword.t()) ::
+  @spec ensure_within_plan_limits(map(), struct() | atom() | nil, Keyword.t()) ::
           :ok | {:error, Limits.over_limits_error()}
-  def ensure_within_plan_limits(user_or_usage, plan, opts \\ [])
 
-  def ensure_within_plan_limits(%User{} = user, %plan_mod{} = plan, opts)
-      when plan_mod in [Plan, EnterprisePlan] do
-    ensure_within_plan_limits(Usage.usage(user), plan, opts)
-  end
+  def ensure_within_plan_limits(usage, plan_mod, opts \\ [])
 
   def ensure_within_plan_limits(usage, %plan_mod{} = plan, opts)
       when plan_mod in [Plan, EnterprisePlan] do
